requirements.txt
Flask
Flask-Login
Flask-Bcrypt
Flask-WTF
Flask-SQLAlchemy
python-dotenv
stripe
qrcode
email-validator
pytz
